# [Draft] Implement Validation pass for train (2nd version)