--- a/runtime/onert/api/include/nnfw_experimental.h
+++ b/runtime/onert/api/include/nnfw_experimental.h
@@ -264,9 +264,31 @@ NNFW_STATUS nnfw_train_set_expected(nnfw_session *session, uint32_t index, const
                                     const nnfw_tensorinfo *expected_info);
 
 /**
+ * @brief Set training output buffer
+ *
+ * This function must be called after {@link nnfw_train_prepare}, \p buffer given to this function
+ * can be reused for training. \p length must be greater or equal than the operand requires.
+ * An output operand can have unspecified shape and deduced dynamically during the execution. You
+ * must provide \p buffer large enough.
+ *
+ * @param[in]   session Session from inference output is to be extracted
+ * @param[in]   index   Index of output to be set (0-indexed)
+ * @param[in]   type    Type of the output
+ * @param[out]  buffer  Raw buffer for output
+ * @param[in]   length  Size of bytes of output buffer
+ *
+ * @return      @c NNFW_STATUS_NO_ERROR if successful
+ */
+NNFW_STATUS nnfw_train_set_output(nnfw_session *session, uint32_t index, NNFW_TYPE type,
+                                  void *buffer, size_t length);
+
+/**
  * @brief Train the model
  * @note  This function should be called after {@link nnfw_train_set_input} and
  *        {@link nnfw_train_set_expected} for each input and expected output
+ *
+ *        In order to use \p update_weights as false, it should be called after
+ *        {@link nnfw_train_set_output}.
  *
  * @param[in] session The session to be trained
  * @param[in] update_weights If true, update weights of the model