--- a/tests/tools/onert_train/src/args.h
+++ b/tests/tools/onert_train/src/args.h
@@ -59,6 +59,7 @@ public:
   const int getLossType(void) const { return _loss_type; }
   const int getLossReductionType(void) const { return _loss_reduction_type; }
   const int getOptimizerType(void) const { return _optimizer_type; }
+  const int getMetricType(void) const { return _metric_type; }
   const bool printVersion(void) const { return _print_version; }
   const int getVerboseLevel(void) const { return _verbose_level; }
   std::unordered_map<uint32_t, uint32_t> getOutputSizes(void) const { return _output_sizes; }
@@ -84,6 +85,7 @@ private:
   int _loss_type;
   int _loss_reduction_type;
   int _optimizer_type;
+  int _metric_type;
   bool _print_version = false;
   int _verbose_level;
   std::unordered_map<uint32_t, uint32_t> _output_sizes;