… tests/tools/onert_train/CMakeLists.txt
… tests/tools/onert_train/src/rawdataloader.h
… tests/tools/onert_train/test/rawdataloader.test.cc