Related issue: #12204