@jyoungyun 2023-12-12
Without `metric` argument:

```
$ ./Product/out/bin/onert_train --modelfile ../Personal/models/mnist/mnist_keras/mnist_keras_softmax2/mnist_keras.circle --load_input:raw ../Personal/models/data/fashion_mnist/train.input.1000.bin --load_expected:raw ../Personal/models/data/fashion_mnist/train.output.1000.bin --loss 1 --optimizer 1 --epoch 5 --batch_size 1000 --learning_rate 0.001 --loss_reduction_type 1
Model Expected Filename ../Personal/models/data/fashion_mnist/train.output.1000.bin
Model Input Filename ../Personal/models/data/fashion_mnist/train.input.1000.bin
Model Filename ../Personal/models/mnist/mnist_keras/mnist_keras_softmax2/mnist_keras.circle
Epoch 1/5 - time: 86.222ms/step - loss: [0] 0.0117 - val_loss: [0] 0.0114
Epoch 2/5 - time: 83.088ms/step - loss: [0] 0.0114 - val_loss: [0] 0.0111
Epoch 3/5 - time: 83.351ms/step - loss: [0] 0.0111 - val_loss: [0] 0.0109
Epoch 4/5 - time: 83.146ms/step - loss: [0] 0.0109 - val_loss: [0] 0.0106
Epoch 5/5 - time: 83.444ms/step - loss: [0] 0.0106 - val_loss: [0] 0.0103
===================================
MODEL_LOAD   takes 0.9820 ms
PREPARE      takes 14.9660 ms
EXECUTE      takes 550.5860 ms
- Epoch 1      takes 86.2220 ms
- Epoch 2      takes 83.0880 ms
- Epoch 3      takes 83.3510 ms
- Epoch 4      takes 83.1460 ms
- Epoch 5      takes 83.4440 ms
===================================
```

With `metric` argument:

```
$ ./Product/out/bin/onert_train --modelfile ../Personal/models/mnist/mnist_keras/mnist_keras_softmax2/mnist_keras.circle --load_input:raw ../Personal/models/data/fashion_mnist/train.input.1000.bin --load_expected:raw ../Personal/models/data/fashion_mnist/train.output.1000.bin --loss 1 --optimizer 1 --epoch 5 --batch_size 1000 --learning_rate 0.001 --loss_reduction_type 1 --metric 0
Model Expected Filename ../Personal/models/data/fashion_mnist/train.output.1000.bin
Model Input Filename ../Personal/models/data/fashion_mnist/train.input.1000.bin
Model Filename ../Personal/models/mnist/mnist_keras/mnist_keras_softmax2/mnist_keras.circle
Epoch 1/5 - time: 85.829ms/step - loss: [0] 0.0117 - val_loss: [0] 0.0114 - val_accuracy: [0] 0.1110
Epoch 2/5 - time: 83.381ms/step - loss: [0] 0.0114 - val_loss: [0] 0.0111 - val_accuracy: [0] 0.1110
Epoch 3/5 - time: 83.136ms/step - loss: [0] 0.0111 - val_loss: [0] 0.0109 - val_accuracy: [0] 0.1100
Epoch 4/5 - time: 83.035ms/step - loss: [0] 0.0109 - val_loss: [0] 0.0106 - val_accuracy: [0] 0.1110
Epoch 5/5 - time: 83.434ms/step - loss: [0] 0.0106 - val_loss: [0] 0.0103 - val_accuracy: [0] 0.1050
===================================
MODEL_LOAD   takes 1.0020 ms
PREPARE      takes 13.4180 ms
EXECUTE      takes 550.7950 ms
- Epoch 1      takes 85.8290 ms
- Epoch 2      takes 83.3810 ms
- Epoch 3      takes 83.1360 ms
- Epoch 4      takes 83.0350 ms
- Epoch 5      takes 83.4340 ms
===================================
```